reworked trade confirmation a little: moved logic to service, split it into methods,secured: user now can confirm only his own trades
